Fixed bug

+++ src/fcnet.py
@@ -21,7 +21,7 @@ def random_init(n_in, n_out, weight_scale=5e-2, dtype=np.float32):
     #                           BEGIN OF YOUR CODE                            #
     ###########################################################################
 
-    W = np.random.normal(0, weight_scale, [n_in, n_out])
+    W = np.random.normal(0, weight_scale, [n_in, n_out]).astype(dtype)
     b = np.zeros(n_out)
 
     ###########################################################################
